requirements.txt
opencv-python
pytesseract
Pillow
keyboard
accessible_output2
pyautogui